feat: TOP-53 resizeDialogFragment 확장 함수로 변경
