+++ core/ui/src/main/java/tht/core/ui/extension/DialogFragmentExtension.kt
@@ -1,0 +1,33 @@
+package tht.core.ui.extension
+
+import android.content.Context
+import android.graphics.Point
+import android.os.Build
+import android.view.WindowManager
+import androidx.fragment.app.DialogFragment
+
+fun DialogFragment.resizeDialogFragment(width: Float, height: Float) {
+    val windowManager = requireContext().getSystemService(Context.WINDOW_SERVICE) as WindowManager
+
+    if (Build.VERSION.SDK_INT < 30) {
+        val display = windowManager.defaultDisplay
+        val size = Point()
+
+        display.getSize(size)
+
+        val window = dialog?.window
+
+        val x = (size.x * width).toInt()
+        val y = (size.y * height).toInt()
+        window?.setLayout(x, y)
+    } else {
+        val rect = windowManager.currentWindowMetrics.bounds
+
+        val window = dialog?.window
+
+        val x = (rect.width() * width).toInt()
+        val y = (rect.height() * height).toInt()
+
+        window?.setLayout(x, y)
+    }
+}
