+++ feature/signin/src/main/java/tht/feature/signin/signup/location/LocationDialogFragment.kt
@@ -20,6 +20,7 @@ import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
 import tht.core.ui.delegate.viewBinding
+import tht.core.ui.extension.resizeDialogFragment
 import tht.feature.signin.databinding.DialogLocationBinding
 
 @AndroidEntryPoint
@@ -29,7 +30,7 @@ class LocationDialogFragment : DialogFragment() {
 
     override fun onStart() {
         super.onStart()
-        resizeDialogFragment(requireContext(), this, 0.9f, 0.5f)
+        resizeDialogFragment(0.9f, 0.5f)
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
@@ -69,32 +70,6 @@ class LocationDialogFragment : DialogFragment() {
                 navController.previousBackStackEntry?.savedStateHandle?.set(LocationConstant.KEY, address)
                 navController.popBackStack()
             }
-        }
-    }
-
-    private fun resizeDialogFragment(context: Context, dialogFragment: DialogFragment, width: Float, height: Float) {
-        val windowManager = context.getSystemService(Context.WINDOW_SERVICE) as WindowManager
-
-        if (Build.VERSION.SDK_INT < 30) {
-            val display = windowManager.defaultDisplay
-            val size = Point()
-
-            display.getSize(size)
-
-            val window = dialogFragment.dialog?.window
-
-            val x = (size.x * width).toInt()
-            val y = (size.y * height).toInt()
-            window?.setLayout(x, y)
-        } else {
-            val rect = windowManager.currentWindowMetrics.bounds
-
-            val window = dialogFragment.dialog?.window
-
-            val x = (rect.width() * width).toInt()
-            val y = (rect.height() * height).toInt()
-
-            window?.setLayout(x, y)
         }
     }
 }
